requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn
python-dotenv
pandas
Jinja2==3.0.3
google-cloud-bigquery
db-dtypes
google-cloud-storage
